Quick hack to show gltf import. Update to React 18
I had some issues with the gltf import that were solved when I added an
Environment from @react-three/drei, but I was getting npm errors when
installing that so I had to update to React 18 and new react three fiber
version.

It could be that we just need to import an older version of
@react-three/drei, or that we just need to update React and we don't
need an Environment. However it would make sense if we need an
Environment as the Sponza gltf scene probably makes use of a reflection
environment for reflective materials.

This is a quick hack, will need to be refactored a lot before merging at
a later date.